Uptimize object update during reconciliation
